Task 6
eddjh
JJLFUF;IU;AA;
exit

+++ 0x08-recursion/6-is_prime_number.c
@@ -7,7 +7,7 @@
  */
 int is_prime_number(int n)
 {
-	unsigned long i = 2;
+	unsigned int i = 2;
 
 	if (n == 0 || n <= 1)
 		return (0);
